Moved get_outmail_state() to its own file
